Use double instead of float for square length, wrote crash tests to verify lots of different graphs and traversals don't cause crashes.

+++ include/utils.hpp
@@ -2,5 +2,5 @@
 
 #include <raylib.h>
 
-float square(float x);
-float distanceSquared(Vector2 v1, Vector2 v2);
+double square(double x);
+double distanceSquared(Vector2 v1, Vector2 v2);
